Don't update the chunkTimes via an update
Because WebSockets can be a little bit flaky, and race conditions can happen, I've removed the functionality that updates the chunk times which otherwise would have been useful to avoid reloading chunks on refreshing the page. But it doesn't work because using this system, it's possible for the state to gradually fall out of sync with reality. The downside of this change is that if Setzen has a high number of users, it's unlikely that state will be persisted well enough that would reduce number of requests of state.